Text-Spelling/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.7.0
baron==0.10.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Counter==1.0.0
crispy-bootstrap5==0.6
cryptography==39.0.0
dill==0.3.6
distlib==0.3.4
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.1
django-cache-url==3.4.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-environs==0.1.1
django-js-asset==2.0.0
djangorestframework==3.14.0
environs==9.5.0
filelock==3.7.1
Flask==2.2.2
frozenlist==1.3.3
googlemaps==4.7.3
googletrans==3.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html-to-json==2.0.0
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==4.9.2
MarkupSafe==2.1.1
marshmallow==3.18.0
MouseInfo==0.1.3
multidict==6.0.4
multiprocess==0.70.14
openai==0.26.0
packaging==21.3
pathos==0.3.0
phonenumbers==8.13.3
Pillow==9.2.0
platformdirs==2.5.2
pox==0.3.2
ppft==1.7.6.6
psycopg2-binary==2.9.3
PyAutoGUI==0.9.53
pycodestyle==2.9.1
pycparser==2.21
PyExecJS==1.5.1
PyGetWindow==0.0.9
PyJWT==2.6.0
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pyTelegramBotAPI==4.7.0
python-decouple==3.6
python-dotenv==0.21.0
pytube==12.1.0
pytweening==1.0.4
pytz==2022.2.1
pywhatkit==5.4
redbaron==0.9.2
requests==2.28.1
rfc3986==1.5.0
rply==0.7.8
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
toml==0.10.2
tqdm==4.64.1
translate==3.6.1
translators==5.5.5
twilio==7.16.0
tzdata==2022.4
Unipath==1.1
urllib3==1.26.13
virtualenv==20.15.1
virtualenv-clone==0.5.7
Werkzeug==2.2.2
whitenoise==6.2.0
wikipedia==1.4.0
yarl==1.8.2
